+++ noodles-bcf/src/async.rs
@@ -1,3 +1,4 @@
 //! Async BCF.
 
+pub mod fs;
 pub mod io;